requirements.txt
asgiref==3.11.1
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
Django==5.1.3
django-ratelimit==4.1.0
django-allauth==65.11.2
pdfminer.six==20260107
pillow==12.1.1
pycparser==3.0
pyotp==2.9.0
python-decouple==3.8
qrcode==8.2
requests>=2.31,<3
PyJWT>=2.9,<3
sqlparse==0.5.5
argon2-cffi>=23.1.0
